+++ packages/ethers-v5/.fatherrc.ts
@@ -1,0 +1,5 @@
+import { defineConfig } from 'father';
+
+export default defineConfig({
+  extends: '../../.fatherrc.base.ts',
+});
