+++ packages/ethers-v5/README.md
@@ -1,0 +1,3 @@
+# @ant-design/web3-ethers
+
+[https://web3.ant.design](https://web3.ant.design)
