+++ packages/ethers-v5/src/global.d.ts
@@ -1,0 +1,5 @@
+import type { EIP1193Provider } from 'viem';
+
+declare interface Window {
+  ethereum?: EIP1193Provider;
+}
