+++ packages/ethers-v5/src/hooks/use-ethers-provider.test.tsx
@@ -1,9 +1,9 @@
 import { render } from '@testing-library/react';
-import { providers } from 'ethers-v5';
+import { providers } from 'ethers';
 import { describe, expect, test } from 'vitest';
 
-import { EthersWeb3ConfigProvider } from '../ethers-provider';
-import { useEthersProvider } from './use-ethers-provider-v5';
+import { EthersWeb3ConfigProvider } from '../provider';
+import { useEthersProvider } from './use-ethers-provider';
 
 describe('useEthersProvider(v5)', async () => {
   test('basic usage', async () => {
