+++ packages/ethers-v5/src/index.ts
@@ -1,0 +1,3 @@
+export * from './provider';
+export * from './wallets';
+export * from './hooks';
