+++ packages/ethers-v5/src/provider.ts
@@ -1,0 +1,1 @@
+export * from '@ant-design/web3-ethers/provider';
