--- packages/ethers/src/hooks/index-v5.ts
@@ -1,2 +1,0 @@
-export * from './use-ethers-provider-v5';
-export * from './use-ethers-signer-v5';
