--- packages/ethers/src/index-v5.ts
@@ -1,3 +1,0 @@
-export * from './ethers-provider';
-export * from './wallets';
-export * from './hooks/index-v5';
