+++ tsconfig.base.json
@@ -22,8 +22,11 @@
       "@ant-design/web3-assets/solana": ["./packages/assets/src/solana/"],
       "@ant-design/web3-wagmi": ["./packages/wagmi/src/"],
       "@ant-design/web3-ethers": ["./packages/ethers/src/"],
-      "@ant-design/web3-ethers/legacy-v5": ["./packages/ethers/src/index-v5.ts"],
+      "@ant-design/web3-ethers/provider": ["./packages/ethers/src/ethers-provider/index.ts"],
+      "@ant-design/web3-ethers/wallets": ["./packages/ethers/src/wallets/index.ts"],
       "@ant-design/web3-ethers/wagmi": ["./packages/ethers/src/wagmi.ts"],
+      "@ant-design/web3-ethers-v5": ["./packages/ethers-v5/src/index.ts"],
+      "@ant-design/web3-ethers-v5/wagmi": ["./packages/ethers-v5/src/wagmi.ts"],
       "@ant-design/web3-solana": ["./packages/solana/src/"]
     }
   }
